Draft SIMD DynamicBloom
Summary: Adds a faster SIMD-based Bloom filter implementation for
the memtable (DynamicBloom) that is selected at compile time whenever
AVX2 support is enabled in the build. The existing USE_SSE flag is
updated to enable building with AVX2 support, in addition to SSE4.2 and
PCLMUL, because it should be available on almost all production
Intel machines, first appearing in Westmere processors 5+ years ago.

This implementation has a similar false positive rate

Test Plan: