Make filter_policy a mutable option, fix race
Summary: I wanted to make filter_policy a mutable option, but the
SetOptions race fix proposed in #10480 doesn't work for object options
like filter_policy. (Race still reported.)

I found that when the options objects are copied for persisting to file,
the table factory (and thus options) is not copied, therefore leading to
a race between persisting the options and subsequent updates with
SetOptions(). It appears that fixing this fixes the race that was being
reported with SetOptions() in the

Fixes #10079

Intended follow-up: See #10021

Test Plan: unit test added, crash test updated with more SetOptions and
higher probability of calling it.

Ran TSAN blackbox crashtest for hours with options to amplify potential
race (see #10079)